Complete Commit 26/1/2023

+++ JAVA WS/Term 2/Programs/Q20.java
@@ -5,8 +5,8 @@ public class Q20 {
     public static void main(String[] args) {
         Scanner in = new Scanner(System.in);
         System.out.print("Enter 2 numbers : ");
-        int a = in.nextInt();
-        int b = in.nextInt();
+        float a = in.nextFloat();
+        float b = in.nextFloat();
         float ans = ((a * b) + (a - b)) / 2;
         System.out.print("((" + a + " * " + b + ") + (" + a + " - " + b + ")) / 2 = " + ans);
     }
